requirements.txt
torch==2.1.0
torchaudio==2.1.0
tensorboard

soundfile
librosa
pydub

g2p_en
pyworld

munch==4.0.0
tqdm
omegaconf
pyyaml
numpy==1.24.4
matplotlib
pandas
jiwer